@@ -21,12 +21,12 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_secure_storage: ^4.2.1
-  flutter_svg: ^0.22.0
-  flutter_translate: ^3.0.1
-  font_awesome_flutter: ^9.1.0
+  flutter_secure_storage: ^6.0.0
+  flutter_svg: ^1.1.4
+  flutter_translate: ^4.0.3
+  font_awesome_flutter: ^10.2.1
   get_it: ^7.2.0
-  http: ^0.13.3
+  http: ^0.13.5
   intl: ^0.17.0
   path: ^1.8.0
   rxdart: ^0.27.2
