@@ -25,13 +25,7 @@ class OpenFlutterBottomMenu extends StatelessWidget {
         width: 24.0,
         color: colorByIndex(theme, index),
       ),
-      title: Text(
-        title,
-        style: TextStyle(
-          fontSize: 10.0,
-          color: colorByIndex(theme, index),
-        ),
-      ),
+      label: title,
     );
   }
 
